# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: Poss POS API
  version: 1.0.0
  description: Versioned REST API foundation. Business operations endpoints are added as their backend modules are implemented.
servers:
  - url: http://localhost:8000/api/v1
paths:
...
  /auth/register-business:
    post:
      summary: Create a business and owner account
      operationId: registerBusiness
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [business_name, name, email, password, password_confirmation]
              properties:
                business_name: { type: string, maxLength: 160 }
                name: { type: string, maxLength: 120 }
                email: { type: string, format: email }
                password: { type: string, minLength: 12, format: password }
                password_confirmation: { type: string, format: password }
      responses:
        '201':
          description: Business, owner membership, and API token created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Sanctum
  schemas:
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
    Business:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        currency: { type: string, example: KES }
        timezone: { type: string, example: Africa/Nairobi }
...
    RegistrationResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [token, user, business, role]
          properties:
            token: { type: string }
            user: { $ref: '#/components/schemas/User' }
            business: { $ref: '#/components/schemas/Business' }
            role: { type: string, const: owner }
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
              errors: { type: object }
    TooManyRequests:
      description: Rate limit exceeded